+++ commerce/interaction-api/src/main/java/ru/yandex/practicum/commerce/payment/client/PaymentClient.java
@@ -7,6 +7,7 @@ import org.springframework.web.bind.annotation.RequestBody;
 import ru.yandex.practicum.commerce.order.dto.OrderDto;
 import ru.yandex.practicum.commerce.payment.dto.PaymentDto;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 @FeignClient(name = "payment", path = "/api/v1/payment")
@@ -16,13 +17,13 @@ public interface PaymentClient {
     PaymentDto createPaymentOrder(@Valid @RequestBody OrderDto orderDto);
 
     @PostMapping("/totalCost")
-    Double calculateTotalCost(@Valid @RequestBody OrderDto orderDto);
+    BigDecimal calculateTotalCost(@Valid @RequestBody OrderDto orderDto);
 
     @PostMapping("/refund")
     void refundPayment(@Valid @RequestBody UUID paymentId);
 
     @PostMapping("/productCost")
-    Double calculateProductCost(@Valid @RequestBody OrderDto orderDto);
+    BigDecimal calculateProductCost(@Valid @RequestBody OrderDto orderDto);
 
     @PostMapping("/failed")
     void setPaymentFailed(@Valid @RequestBody UUID paymentId);
