+++ commerce/interaction-api/src/main/java/ru/yandex/practicum/commerce/payment/dto/PaymentDto.java
@@ -4,15 +4,16 @@ import lombok.Getter;
 import lombok.Setter;
 import ru.yandex.practicum.commerce.payment.enums.PaymentStatus;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 @Getter
 @Setter
 public class PaymentDto {
     private UUID paymentId;
-    private Double totalPayment;
-    private Double deliveryTotal;
-    private Double feeTotal;
-    private Double productTotal;
+    private BigDecimal totalPayment;
+    private BigDecimal deliveryTotal;
+    private BigDecimal feeTotal;
+    private BigDecimal productTotal;
     private PaymentStatus status;
 }
