+++ commerce/order/src/main/java/ru/yandex/practicum/commerce/order/service/OrderService.java
@@ -19,6 +19,7 @@ import ru.yandex.practicum.commerce.warehouse.controller.WarehouseClient;
 import ru.yandex.practicum.commerce.warehouse.entity.dto.AssemblyProductsForOrderRequest;
 import ru.yandex.practicum.commerce.warehouse.entity.dto.BookedProductsDto;
 
+import java.math.BigDecimal;
 import java.util.List;
 import java.util.UUID;
 
@@ -72,17 +73,17 @@ public class OrderService {
         Order order = findOrderById(orderId);
         order.setState(OrderState.ON_PAYMENT);
         if (order.getProductPrice() == null) {
-            Double productCost = paymentClient.calculateProductCost(orderMapper.toDto(order));
+            BigDecimal  productCost = paymentClient.calculateProductCost(orderMapper.toDto(order));
             order.setProductPrice(productCost);
         }
 
         if (order.getDeliveryPrice() == null) {
-            Double deliveryCost = deliveryClient.getDeliveryCost(orderMapper.toDto(order));
+            BigDecimal  deliveryCost = deliveryClient.getDeliveryCost(orderMapper.toDto(order));
             order.setDeliveryPrice(deliveryCost);
         }
 
         if (order.getTotalPrice() == null) {
-            Double totalCost = paymentClient.calculateTotalCost(orderMapper.toDto(order));
+            BigDecimal  totalCost = paymentClient.calculateTotalCost(orderMapper.toDto(order));
             order.setTotalPrice(totalCost);
         }
         OrderDto orderDto = orderMapper.toDto(order);
@@ -133,14 +134,14 @@ public class OrderService {
     public OrderDto calculateTotal(UUID orderId) {
         Order order = findOrderById(orderId);
         if (order.getProductPrice() == null) {
-            Double productCost = paymentClient.calculateProductCost(orderMapper.toDto(order));
+            BigDecimal productCost = paymentClient.calculateProductCost(orderMapper.toDto(order));
             order.setProductPrice(productCost);
         }
         if (order.getDeliveryPrice() == null) {
-            Double deliveryCost = deliveryClient.getDeliveryCost(orderMapper.toDto(order));
+            BigDecimal  deliveryCost = deliveryClient.getDeliveryCost(orderMapper.toDto(order));
             order.setDeliveryPrice(deliveryCost);
         }
-        Double total = paymentClient.calculateTotalCost(orderMapper.toDto(order));
+        BigDecimal  total = paymentClient.calculateTotalCost(orderMapper.toDto(order));
         order.setTotalPrice(total);
         orderRepository.save(order);
         return orderMapper.toDto(order);
@@ -148,7 +149,7 @@ public class OrderService {
 
     public OrderDto calculateDelivery(UUID orderId) {
         Order order = findOrderById(orderId);
-        Double deliveryTotal = deliveryClient.getDeliveryCost(orderMapper.toDto(order));
+        BigDecimal  deliveryTotal = deliveryClient.getDeliveryCost(orderMapper.toDto(order));
         order.setDeliveryPrice(deliveryTotal);
         orderRepository.save(order);
         return orderMapper.toDto(order);
